fixed: skip if alreay exsit

+++ install/install_shell.sh
@@ -14,13 +14,17 @@ sudo apt-get install -y \
   curl \
   bat
 
-toilet -w 200 -f future "Installing: Nerd font FiraCode"
-mkdir -p ~/.local/share/fonts
-cd ~/.local/share/fonts
-curl -fLO https://github.com/ryanoasis/nerd-fonts/releases/latest/download/FiraCode.zip
-unzip FiraCode.zip
-rm FiraCode.zip
-fc-cache -fv
+if fc-list | grep -qi "FiraCode"; then
+  toilet -w 20 -f future "Skipping: Nerd font FiraCode (already installed)"
+else
+  toilet -w 200 -f future "Installing: Nerd font FiraCode"
+  mkdir -p "$HOME"/.local/share/fonts
+  cd "$HOME"/.local/share/fonts
+  curl -fLO https://github.com/ryanoasis/nerd-fonts/releases/latest/download/FiraCode.zip
+  unzip FiraCode.zip
+  rm FiraCode.zip
+  fc-cache -fv
+fi
 
 cd $PATH_SCRIPT
 
@@ -32,10 +36,22 @@ sudo apt-get install -y tmux
 
 toilet -w 200 -f future "Installing: Zsh advanced shell and its plugins"
 sudo apt-get install -y zsh
-sh -c "$(curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh)"
+if [ -d "$HOME/.oh-my-zsh" ]; then
+  toilet -w 200 -f future "Skipping: Oh-My-Zsh (already installed)"
+else
+  sh -c "$(curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh)"
+fi
+
 # setup zsh as default shell
-chsh -s $(which zsh)
+if [ "$SHELL" != "$(which zsh)" ]; then
+  chsh -s $(which zsh)
+fi
 
 toilet -w 200 -f future "Installing: Starship shell prompt"
-curl -sS https://starship.rs/install.sh | sh
+if [ ! command -v starship ] &>/dev/null; then
+  toilet -w 200 -f future "Installing: Starship"
+  curl -sS https://starship.rs/install.sh | sh
+else
+  toilet -w 200 -f future "Skipping: Starship (already installed)"
+fi
 exit 0
